[Fix] User add routing

+++ source/server/routes/rest/user/user.js
@@ -170,25 +170,28 @@ async function upsertUser(req, res){
         }
       }); 
     }
-
-    await models.sequelize.query(`DELETE FROM fielduser where uid = '${req.body.uid}'`);
+    console.log(req.body);
+    await models.sequelize.query(`DELETE FROM fielduser where userUuid = '${req.body.uid}'`);
     //add field
-    await req.body.fields.forEach(async (f) => {
-      try{
-        await models.sequelize.query(`INSERT INTO fielduser (field, userUid, createdAt, updatedAt) VALUES ('${f}','${req.body.uid}',CURRENT_TIME, CURRENT_TIME)`);
+    if(req.body.fields.length > 0){
+      await req.body.fields.forEach(async (f) => {
+        try{
+          await models.sequelize.query(`INSERT INTO fielduser (field, userUid, createdAt, updatedAt) VALUES ('${f}','${req.body.uid}',CURRENT_TIME, CURRENT_TIME)`);
 
-      } catch (err){
-        
-      }
-    });
-    if(type == "Mentor"){
+        } catch (err){
+
+        }
+      });
+    }
+
+    if(type == "Mentor" && req.body.available_times.length > 0){
       await models.sequelize.query(`DELETE FROM timeuser where uid = '${req.body.uid}'`);
       //add field
       await req.body.available_times.forEach(async (t) => {
         try{
           await models.sequelize.query(`INSERT INTO timeuser (uid, timeId, createdAt, updatedAt) VALUES ('${req.body.uid}','${t}',CURRENT_TIME, CURRENT_TIME)`);
         } catch (err){
-          
+
         }
       });
     }
